Fixes pagination bug
Pagination stops being shown when the user is on the last page.  What
should happen is that pagination isn't shown when there's only one page
of data.
