+++ src/views/partials/pagination.ejs
@@ -1,0 +1,80 @@
+<div class="govuk-grid-row">
+    <div class="govuk-grid-column-two-thirds">
+        <nav class="govuk-pagination" role="navigation" aria-label="Pagination">
+            <% if (locals.total_pages === 1) { %>
+                <ul class="govuk-pagination__list">
+                    <li class="govuk-pagination__item govuk-pagination__item--active">
+                                    <span aria-current="page" class="govuk-pagination__link">
+                                        <%= locals.current_page %>
+                                    </span>
+                    </li>
+                </ul>
+            <% } else { %>
+                <% if (locals.current_page > 1) { %>
+                    <div class="govuk-pagination__prev">
+                        <a class="govuk-link govuk-pagination__link"
+                           href="<%= buildUrl(`/publish/${locals.datasetId}/preview`, i18n.language, { page_number: locals.current_page - 1, page_size: locals.page_size }) %>"
+                           rel="prev">
+                            <svg class="govuk-pagination__icon govuk-pagination__icon--prev" xmlns="http://www.w3.org/2000/svg"
+                                 height="13" width="15" aria-hidden="true" focusable="false" viewBox="0 0 15 13">
+                                <path
+                                        d="m6.5938-0.0078125-6.7266 6.7266 6.7441 6.4062 1.377-1.449-4.1856-3.9768h12.896v-2h-12.984l4.2931-4.293-1.414-1.414z">
+                                </path>
+                            </svg>
+                            <span class="govuk-pagination__link-title">
+                                        <%= t('pagination.previous') %>
+                                    </span>
+                        </a>
+                    </div>
+                <% } %>
+                <ul class="govuk-pagination__list">
+                    <% locals.pagination.forEach((item) => { %>
+                        <% if (item === '...') { %>
+                            <li class="govuk-pagination__item govuk-pagination__item--ellipses">
+                                &ctdot;
+                            </li>
+                        <% } else if(item === locals.current_page) { %>
+                            <li class="govuk-pagination__item govuk-pagination__item--active">
+                                            <span aria-current="page" class="govuk-pagination__link">
+                                                <%= item %>
+                                            </span>
+                            </li>
+                        <% } else { %>
+                            <li class="govuk-pagination__item">
+                                <a class="govuk-link govuk-pagination__link"
+                                   href="<%= buildUrl(`/publish/${locals.datasetId}/preview`, i18n.language, { page_number: item, page_size: locals.page_size }) %>"
+                                   aria-label="Page <%= item %>">
+                                    <%= item %>
+                                </a>
+                            </li>
+                        <% }%>
+                    <% }); %>
+                </ul>
+                <% if (locals.current_page < locals.total_pages) { %>
+                    <div class="govuk-pagination__next">
+                        <a class="govuk-link govuk-pagination__link"
+                           href="<%= buildUrl(`/publish/${locals.datasetId}/preview`, i18n.language, { page_number: locals.current_page + 1, page_size: locals.page_size }) %>"
+                           rel="next">
+                                    <span class="govuk-pagination__link-title">
+                                        <%= t('pagination.next') %>
+                                    </span>
+                            <svg class="govuk-pagination__icon govuk-pagination__icon--next"
+                                 xmlns="http://www.w3.org/2000/svg" height="13" width="15" aria-hidden="true"
+                                 focusable="false" viewBox="0 0 15 13">
+                                <path
+                                        d="m8.107-0.0078125-1.4136 1.414 4.2926 4.293h-12.986v2h12.896l-4.1855 3.9766 1.377 1.4492 6.7441-6.4062-6.7246-6.7266z">
+                                </path>
+                            </svg>
+                        </a>
+                    </div>
+                <% } %>
+            <% } %>
+        </nav>
+    </div>
+
+    <div class="govuk-grid-column-one-third">
+        <p class="govuk-body" style="text-align: right; padding-top: 10px">
+            <%= t('publish.preview.showing_rows', {start: locals.page_info.start_record, end: locals.page_info.end_record, total: locals.page_info.total_records}) %>
+        </p>
+    </div>
+</div>
